move shared code out of scenes
