+++ .flowconfig
@@ -11,6 +11,14 @@
 
 [options]
 emoji=true
+
 suppress_type=$FlowIssue
 suppress_type=$FlowFixMe
 suppress_type=$FixMe
+
+; we'd like to do this, but it doesn't seem to work:
+; module.name_mapper='^shared/(.*)$' -> '<PROJECT_ROOT>/src/shared/\1'
+; https://github.com/facebook/flow/issues/1068
+; use the entire src directory as a resolve root in the meantime
+module.system.node.resolve_dirname=node_modules
+module.system.node.resolve_dirname=src
