+++ src/scenes/bills/components/loading_indicator.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react'
 import { StyleSheet, css } from 'aphrodite/no-important'
 import BillAnimation, { billStyle } from './bill_animation'
-import { colors, shadows, borders } from '../../styles'
+import { colors, shadows, borders } from 'shared/styles'
 
 export default class LoadingIndicator extends Component {
   props: {
