+++ src/shared/components/__tests__/icon_link.test.js
