+++ src/shared/components/index.js
