+++ src/shared/styles/colors.js
