+++ src/shared/types/__tests__/connection.test.js
