+++ src/shared/types/connection.js
