+++ src/shared/types/edge.js
