+++ src/shared/types/page_info.js
